requirements.txt
astronomer-cosmos==1.7.0
apache-airflow-providers-snowflake==5.3.1
apache-airflow-providers-http==4.10.0
snowflake-connector-python==3.10.0
snowflake-sqlalchemy==1.6.1
faker==24.0.0
pandas==2.2.0
requests==2.31.0
great_expectations==0.18.15
python-dotenv==0.21.0